OHH Were Half Way There...
This works on both sides now. It is capable of double checking for the skystone on both sides now.  My only concern is that I did not test what happens when the skystone is picked up normally - perhaps we might need to lower the color sensor value when the second check happens. We might want to check that before we do any merging.