Made curve steeper and shifted graph depending on foundation clamp
When the foundation is clamped, the mass the motors have to move increases, so the minimum threshold of force required to start moving it increases. So, when the foundation is clamped, the exponential graph is shifted up to account for this. Also, i made the graph even steeper to give drivers even more control at lower speeds