Automatic foundation move (WIP)
This basically assigns a button on the controller to move the foundation. When the b key is pressed, the robot will move backwards at a slow speed for 1 second. Hopefully with testing the optimal speed that leave the tower standing 100% of the time will be found. For now, the speed is set to half of max speed.